test: 호감상대 삭제 테스트케이스 추가

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/controller/LikeablePersonController.java
@@ -62,7 +62,7 @@ public class LikeablePersonController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Integer id) {
+    public String delete(@PathVariable("id") Long id) {
 
         LikeablePerson likeablePerson = likeablePersonService.getLikeablePerson(id);
